Create README.md
A COVID-19 tracker is a mobile application or website that provides real-time information and updates on the COVID-19 pandemic. These trackers gather data from reliable sources, such as government health agencies and international organizations, and present it in a user-friendly format.

The primary purpose of a COVID-19 tracker is to keep users informed about the latest statistics and developments related to the virus. This includes the number of confirmed cases, active cases, recoveries, and deaths in different regions or countries in the World.

COVID-19 trackers have become essential tools for individuals, healthcare professionals, and policymakers alike. They help to raise awareness, monitor the spread of the virus, track the effectiveness of interventions and measures, and assist in making informed decisions regarding public health and safety.